test(dynamic-routes): added `useParser` method, and changed syntax for using dynamic-routes
