+++ dynamic-routes/next-10/package.json
@@ -1,3 +1,3 @@
 {
-  "main": "../../dist/dynamic-routes/extensions-10"
+  "main": "../../dist/dynamic-routes/next-10"
 }
