+++ dynamic-routes/next-11/package.json
@@ -1,3 +1,3 @@
 {
-  "main": "../../dist/dynamic-routes/extensions-11.1"
+  "main": "../../dist/dynamic-routes/next-11"
 }
