+++ dynamic-routes/next-12/package.json
@@ -1,3 +1,3 @@
 {
-  "main": "../../dist/dynamic-routes/extensions-12"
+  "main": "../../dist/dynamic-routes/next-12"
 }
