+++ src/dynamic-routes/next-11.ts
@@ -8,9 +8,9 @@ import {
 // @ts-expect-error
 import { normalizePagePath } from "next/dist/server/normalize-page-path";
 
-import { defineRegisterPaths } from "./MemoryRouter.registerPaths";
+import { factory } from "./createDynamicRouteParser";
 
-defineRegisterPaths({
+export const createDynamicRouteParser = factory({
   getSortedRoutes,
   getRouteMatcher,
   getRouteRegex,
