+++ src/dynamic-routes/next-12.ts
@@ -8,9 +8,9 @@ import {
 //
 import { normalizePagePath } from "next/dist/shared/lib/page-path/normalize-page-path";
 
-import { defineRegisterPaths } from "./MemoryRouter.registerPaths";
+import { factory } from "./createDynamicRouteParser";
 
-defineRegisterPaths({
+export const createDynamicRouteParser = factory({
   getSortedRoutes,
   getRouteMatcher,
   getRouteRegex,
